lift ✅: tz1as7NAneT4GhtvQGDZvnXhzAbiCeKedXVP
true to discord, the account that funded this acc was in fact cming, was very open about it and said they will stick to the rules with this account – the one piece on here looks legit, also checked their etherum pieces. altho they look sus it looks like they are not violating copyright as far as i can tell. we should keep this account on the watchlist and check new mints

Lifted by zackundweg