ban 🚫: tz1X3fg4pXCL8k9t6y12m6spzDQxvX3pJGie
mail sent to the original artist (https://www.instagram.com/kinkykarrot_/?hl=de) asking for verification if the account is genuine. this is because we got an automated claim about the acc via mail via google via https://www.ceartas.io/ claiming the acc was not by the original author. mid chance its a false alarm because it was an automated claim and the acc seems very authentic. merch sent a mail to the original artist asking for an answer to verify its an authentic account

Banned by merchantcoppola