lift ✅: tz1X3fg4pXCL8k9t6y12m6spzDQxvX3pJGie
the content notification from google was aoutmated. the artist is not active on tezos but unser crcdng http://crcdng.com/ is their manager and has a contract that allows them to speak on behalf of the artist andf mint their poieces, so if there is any legal action coming in we can refer to crcdng, he will also send us the contracts so we can show them.

Lifted by merchantcoppola